pyproject.toml
[project]
name = "notify-service"
version = "1.0.0"
description = "FastAPI notification service for Telegram and Feishu"
authors = [
    { name = "Notification Service" }
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "loguru>=0.7.2",
    "aiohttp>=3.9.1",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.9"

[project.scripts]
notify-service = "src.main:main"

[tool.hatch.envs.default]
installer = "uv"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true